Refactor database to reuse existing types
