+++ osu.Game/Beatmaps/BeatmapSet.cs
@@ -3,6 +3,7 @@
 
 using System.Collections.Generic;
 using osu.Game.Users;
+using SQLite;
 
 namespace osu.Game.Beatmaps
 {
@@ -11,12 +12,15 @@ namespace osu.Game.Beatmaps
     /// </summary>
     public class BeatmapSet
     {
-        public int BeatmapSetID;
-        
+        [PrimaryKey]
+        public int BeatmapSetID { get; set; }
+        [NotNull, Indexed]
+        public int BeatmapMetadataID { get; set; }
+        [Ignore]
         public List<Beatmap> Beatmaps { get; protected set; }
-
-        public Metadata Metadata;
-
-        public User Creator;
+        [Ignore]
+        public BeatmapMetadata Metadata { get; set; }
+        [Ignore]
+        public User Creator { get; set; }
     }
 }
