--- osu.Game/Database/BeatmapSet.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using SQLite;
-
-namespace osu.Game.Database
-{
-    public class BeatmapSet
-    {
-        [PrimaryKey]
-        public int BeatmapSetID { get; set; }
-        [NotNull, Indexed]
-        public int BeatmapMetadataID { get; set; }
-    }
-}
-
